+++ lib/glimmer/swt/custom/animation.rb
@@ -28,6 +28,7 @@ module Glimmer
       # Represents an animation declaratively
       class Animation
         include Properties
+        include Glimmer::DataBinding::ObservableModel
         
         class << self
           def schedule_frame_animation(animation, &frame_animation_block)
@@ -77,18 +78,19 @@ module Glimmer
           end
         end
         
-        attr_reader :parent, :options, :frame_index, :cycle
+        attr_reader :parent, :options
+        attr_accessor :frame_index, :cycle, :frame_block, :every, :cycle_count, :frame_count, :started, :duration_limit, :duration, :finished, :cycle_count_index
         alias current_frame_index frame_index
-        attr_accessor :frame_block, :every, :cycle_count, :frame_count, :started, :duration_limit
         alias started? started
+        alias finished? finished
         # TODO consider supporting an async: false option
         
         def initialize(parent)
           @parent = parent
           @parent.requires_shape_disposal = true
-          @started = true
-          @frame_index = 0
-          @cycle_count_index = 0
+          self.started = true
+          self.frame_index = 0
+          self.cycle_count_index = 0
           @start_number = 0 # denotes the number of starts (increments on every start)
           self.class.swt_display # ensures initializing variable to set from GUI thread
         end
@@ -111,9 +113,10 @@ module Glimmer
         def start
           return if @start_number > 0 && started?
           @start_number += 1
-          @started = true
           @start_time = Time.now
           @original_start_time = @start_time if @duration.nil?
+          self.finished = false if finished?
+          self.started = true
           # TODO track when finished in a variable for finite animations (whether by frame count, cycle count, or duration limit)
           Thread.new do
             start_number = @start_number
@@ -132,16 +135,16 @@ module Glimmer
         
         def stop
           return if stopped?
-          @started = false
-          @duration = (Time.now - @start_time) + @duration.to_f if duration_limited? && !@start_time.nil?
+          self.started = false
+          self.duration = (Time.now - @start_time) + @duration.to_f if duration_limited? && !@start_time.nil?
         end
         
         # Restarts an animation (whether indefinite or not and whether stopped or not)
         def restart
           @original_start_time = @start_time = nil
-          @duration = nil
-          @frame_index = 0
-          @cycle_count_index = 0
+          self.duration = nil
+          self.frame_index = 0
+          self.cycle_count_index = 0
           stop
           start
         end
@@ -183,6 +186,21 @@ module Glimmer
           end
         end
         
+        def cycle_count_index=(value)
+          @cycle_count_index = value
+          self.finished = true if cycle_limited? && @cycle_count_index == @cycle_count
+        end
+        
+        def frame_index=(value)
+          @frame_index = value
+          self.finished = true if frame_count_limited? && @frame_index == @frame_count
+        end
+        
+        def duration=(value)
+          @duration = value
+          self.finished = true if surpassed_duration_limit?
+        end
+        
         def cycle_enabled?
           @cycle.is_a?(Array)
         end
@@ -196,7 +214,7 @@ module Glimmer
         end
         
         def frame_count_limited?
-          @frame_count.is_a?(Integer)
+          @frame_count.is_a?(Integer) && @frame_count > 0
         end
         
         def surpassed_duration_limit?
@@ -211,11 +229,13 @@ module Glimmer
         
         # Returns true on success of painting a frame and false otherwise
         def draw_frame(start_number)
-          return false if stopped? ||
-                          start_number != @start_number ||
-                          (frame_count_limited? && @frame_index == @frame_count) ||
-                          (cycle_limited? && @cycle_count_index == @cycle_count) ||
-                          surpassed_duration_limit?
+          if stopped? or
+              (start_number != @start_number) or
+              (frame_count_limited? && @frame_index == @frame_count) or
+              (cycle_limited? && @cycle_count_index == @cycle_count) or
+              surpassed_duration_limit?
+            return false
+          end
           block_args = [@frame_index]
           block_args << @cycle[@frame_index % @cycle.length] if cycle_enabled?
           current_frame_index = @frame_index
@@ -231,15 +251,15 @@ module Glimmer
                 @shapes = @parent.shapes - parent_shapes_before
               end
             else
+              self.finished = true if surpassed_duration_limit?
               if stopped? && @frame_index > current_frame_index
-                @started = false
-                @frame_index = current_frame_index
-                @cycle_count_index = current_cycle_count_index
+                self.frame_index = current_frame_index
+                self.cycle_count_index = current_cycle_count_index
               end
             end
           end
-          @frame_index += 1
-          @cycle_count_index += 1 if cycle_limited? && (@frame_index % @cycle&.length&.to_i) == 0
+          self.frame_index += 1
+          self.cycle_count_index += 1 if cycle_limited? && (@frame_index % @cycle&.length&.to_i) == 0
           sleep(every) if every.is_a?(Numeric) # TODO consider using timer_exec as a more reliable alternative
           true
         rescue => e