+++ samples/hello/hello_canvas_animation.rb
@@ -25,31 +25,73 @@ require 'bigdecimal'
 class HelloCanvasAnimation
   include Glimmer::UI::CustomShell
   
-  attr_accessor :delay_time
+  # data-bindable attributes (names must vary from attribute names on animation)
+  attr_accessor :animation_every, :animation_frame_count, :animation_started, :animation_finished
   
   before_body {
-    @delay_time = 0.050
+    @animation_every = 0.050
+    @animation_frame_count = 100
+    @animation_started = true
+    @animation_finished = false
   }
   
   body {
     shell {
-      text 'Hello, Canvas Animation Data Binding!'
-      minimum_size 320, 320
+      grid_layout(2, true)
+      text 'Hello, Canvas Animation!'
+      
+      button {
+        layout_data(:fill, :center, true, false)
+        text <= [self, :animation_started, on_read: ->(value) { value ? 'Stop' : 'Resume' }]
+        enabled <= [self, :animation_finished, on_read: :!]
+        
+        on_widget_selected do
+          if @animation.started?
+            @animation.stop
+          else
+            @animation.start
+          end
+        end
+      }
+      button {
+        layout_data(:fill, :center, true, false)
+        text 'Restart'
+        
+        on_widget_selected do
+          @animation.restart
+        end
+      }
+      label {
+        text 'every (milliseconds)'
+      }
+      label {
+        text 'frame count (0 is unlimited)'
+      }
+      spinner {
+        layout_data(:fill, :center, true, false)
+        digits 3
+        minimum 1
+        maximum 100
+        selection <=> [self, :animation_every, on_read: ->(v) {(BigDecimal(v.to_s)*1000).to_f}, on_write: ->(v) {(BigDecimal(v.to_s)/1000).to_f}]
+      }
+      spinner {
+        layout_data(:fill, :center, true, false)
+        minimum 0
+        maximum 100
+        selection <=> [self, :animation_frame_count]
+      }
       
       canvas {
-        grid_layout
-        
-        spinner {
-          layout_data(:center, :center, true, true) {
-            minimum_width 75
-          }
-          digits 3
-          minimum 1
-          maximum 100
-          selection <=> [self, :delay_time, on_read: ->(v) {(BigDecimal(v.to_s)*1000).to_f}, on_write: ->(v) {(BigDecimal(v.to_s)/1000).to_f}]
+        layout_data(:fill, :fill, true, true) {
+          horizontal_span 2
+          width_hint 320
+          height_hint 320
         }
-        animation {
-          every <= [self, :delay_time]
+        @animation = animation {
+          every <= [self, :animation_every]
+          frame_count <= [self, :animation_frame_count]
+          started <=> [self, :animation_started]
+          finished <=> [self, :animation_finished]
           
           frame { |index|
             background rgb(index%100, index%100 + 100, index%55 + 200)